Ability to keep videos of passed tests (#512)
* added ability to override capabilites from environment variables

* fix checkstyle errors

* fix indentation

* Updated logic how to update capabilities. Now, if we want to update some property, we need to pass this as an environment variable in a specific format:
ex: to update android's app path, set env var as: 'atd.android.app'
ex: To update a specific property as part of an array, 'atd.hostMachines.0.appiumServerPath=.... '
ex: To update the 1st simulator's device name in the 1st hostMachine: 'atd.hostMachines.0.simulators.0.deviceName

* replace . with _ in env variable names

* added TestExecutionContext & SessionContext - and invoke it from the listener

* added loaded Config file also to TestExecutionContext

* Add & remove TestExecutionContext when running methods in parallel as well

* Do NOT delete video of test, even if test is successful IF environment variable - KEEP_ALL_VIDEOS is set to true

* do not pass iTestResult when removing a TestExecutionContext from SessionContext

* remove unnecessary call to getTestExecutionContext

* removed unnecessary sout

* remove sout

+++ src/main/java/com/appium/manager/TestLogger.java
@@ -136,13 +136,14 @@ class TestLogger extends Helpers {
     }
 
     private void deleteSuccessVideos(ITestResult result, String className) {
-        if (result.isSuccess()) {
+        if (result.isSuccess()
+                && (null != System.getenv("KEEP_ALL_VIDEOS"))
+                && !(System.getenv("KEEP_ALL_VIDEOS").equalsIgnoreCase("true"))) {
             File videoFile = new File(System.getProperty("user.dir")
                     + FileLocations.ANDROID_SCREENSHOTS_DIRECTORY
                     + AppiumDeviceManager.getAppiumDevice().getDevice().getUdid() + "/"
                     + className + "/" + result.getMethod().getMethodName()
                     + "/" + result.getMethod().getMethodName() + ".mp4");
-            System.out.println(videoFile);
             if (videoFile.exists()) {
                 videoFile.delete();
             }
